+++ moviepy/Clip.py
@@ -422,6 +422,14 @@ class Clip:
                 end_time = self.duration + end_time
 
         if end_time is not None:
+            # Allow a slight tolerance to account for rounding errors
+            if (self.duration is not None) and (end_time - self.duration > 0.00000001):
+                raise ValueError(
+                    "end_time (%.02f) " % end_time
+                    + "should be smaller or equal to the clip's "
+                    + "duration (%.02f)." % self.duration
+                )
+
             new_clip.duration = end_time - start_time
             new_clip.end = new_clip.start + new_clip.duration