+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support pillow 11
 
 ### Changed <!-- for changes in existing functionality -->
+- Subclipping outside of clip boundaries now raise an exception
 
 ### Deprecated <!-- for soon-to-be removed features -->